# core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.AutomationAuditLogResponse:
    properties:
      action:
        type: string
      apiKeyId:
        type: string
      apiKeyPrefix:
        type: string
      createdAt:
        type: string
      id:
        type: string
      method:
        type: string
      path:
        type: string
      requestBody:
        items:
          type: integer
        type: array
      resourceId:
        type: string
      resourceType:
        type: string
      responseStatus:
        type: integer
      tenantId:
        type: string
      workspaceId:
        type: string
    type: object
...
  github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.AutomationKeyResponse:
    properties:
      allowedTenants:
        items:
          type: string
        type: array
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      keyPrefix:
        type: string
      lastUsedAt:
        type: string
      name:
        type: string
      revokedAt:
        type: string
    type: object
...
  github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.CreateAutomationKeyRequest:
    properties:
      allowedTenants:
        items:
          type: string
        type: array
      name:
        type: string
    required:
    - name
    type: object
  github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.CreateAutomationKeyResponse:
    properties:
      allowedTenants:
        items:
          type: string
        type: array
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      keyPrefix:
        type: string
      name:
        type: string
      rawKey:
        description: shown exactly once
        type: string
    type: object
...
  github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
...
  ? github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ListResponse-github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto_AutomationAuditLogResponse
  : properties:
      count:
        type: integer
      data:
        items:
          $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.AutomationAuditLogResponse'
        type: array
    type: object
  ? github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ListResponse-github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto_AutomationKeyResponse
  : properties:
      count:
        type: integer
      data:
        items:
          $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.AutomationKeyResponse'
        type: array
    type: object
...
  github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.UpdateAutomationKeyRequest:
    properties:
      allowedTenants:
        items:
          type: string
        type: array
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Document Assembly System API - Template management and document generation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Doc Engine API
  version: "1.0"
paths:
  /admin/automation-keys/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ListResponse-github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto_AutomationKeyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List automation API keys
      tags:
      - Automation Keys
    post:
      consumes:
      - application/json
      description: Creates a new API key. The raw key is returned ONLY in this response.
      parameters:
      - description: API key data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.CreateAutomationKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.CreateAutomationKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create automation API key
      tags:
      - Automation Keys
  /admin/automation-keys/{id}:
    delete:
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Revoke automation API key
      tags:
      - Automation Keys
    get:
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.AutomationKeyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get automation API key
      tags:
      - Automation Keys
    patch:
      consumes:
      - application/json
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      - description: Key data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.UpdateAutomationKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.AutomationKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update automation API key
      tags:
      - Automation Keys
  /admin/automation-keys/{id}/audit:
    get:
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      - default: 20
        description: Limit results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset results
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ListResponse-github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto_AutomationAuditLogResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get audit log for API key
      tags:
      - Automation Keys
...
securityDefinitions:
  AutomationKey:
    description: Automation API key for programmatic access
    in: header
    name: X-Automation-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"